refactor: restructure folders and files
